+++ src/main/java/blackjack/view/Formatter.java
@@ -1,7 +1,7 @@
 package blackjack.view;
 
-import blackjack.domain.Participant;
-import blackjack.domain.Player;
+import blackjack.domain.game.Participant;
+import blackjack.domain.game.Player;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardSuit;
